Reorganize table of contents
This draft PR contains some examples of how I believe we can refactor the docs to address the new requirement that ToC nodes can't be expandable and have content at the same time.

It is not complete and not full tried. It is also opinionated :smile: For example, I decided to merge the contents of the what-is-new/roadmap into what-is-new/index and rename it to be the roadmap. I think many customers are looking for a page called the roadmap and this provides that kind of information. On the other hand the what-is-new/index page didn't have enough value to keep it. I think we can follow a similar pattern for the EF6 docs.

Fixes #1517 (well, not yet)